refactor(arch): split the `rp2040` arch into multiple submodule files
Split the `rp2040` arch to make it more manageable when it keeps growing.
