+++ src/riot-rs-embassy/src/arch/rp2040/mod.rs
@@ -1,3 +1,6 @@
+#[cfg(feature = "usb")]
+pub mod usb;
+
 pub(crate) use embassy_executor::InterruptExecutor as Executor;
 pub use embassy_rp::interrupt;
 pub use embassy_rp::interrupt::SWI_IRQ_1 as SWI;
@@ -10,27 +13,6 @@ unsafe fn SWI_IRQ_1() {
     // - not called before `start()`, as the interrupt is enabled by `start()`
     //   itself
     unsafe { crate::EXECUTOR.on_interrupt() }
-}
-
-#[cfg(feature = "usb")]
-pub mod usb {
-    use embassy_rp::{
-        bind_interrupts, peripherals,
-        usb::{Driver, InterruptHandler},
-    };
-
-    use crate::arch;
-
-    bind_interrupts!(struct Irqs {
-        USBCTRL_IRQ => InterruptHandler<peripherals::USB>;
-    });
-
-    pub type UsbDriver = Driver<'static, peripherals::USB>;
-
-    pub fn driver(peripherals: &mut arch::OptionalPeripherals) -> UsbDriver {
-        let usb = peripherals.USB.take().unwrap();
-        Driver::new(usb, Irqs)
-    }
 }
 
 pub fn init(config: Config) -> OptionalPeripherals {
